YoutubeRegistrationRequestのディレクトリ整理
Fixes #611
