+++ src/YoutubeRegistrationRequest/YoutubeRegistrationRequestAcceptingNotification.resolver.ts
@@ -1,10 +1,10 @@
 import { GraphQLError } from "graphql";
 import z from "zod";
 
-import { Resolvers } from "../graphql.js";
-import { resolverNotification } from "../Notification/resolver.js";
-import { ResolverDeps } from "../types.js";
-import { YoutubeRegistrationRequestAcceptingModel } from "../YoutubeRegistrationRequestAccepting/model.js";
+import { Resolvers } from "../resolvers/graphql.js";
+import { resolverNotification } from "../resolvers/Notification/resolver.js";
+import { ResolverDeps } from "../resolvers/types.js";
+import { YoutubeRegistrationRequestAcceptingDTO } from "./dto.js";
 
 export const resolverYoutubeRegistrationRequestAcceptingNotification = ({
   prisma,
@@ -24,7 +24,7 @@ export const resolverYoutubeRegistrationRequestAcceptingNotification = ({
       }
       return prisma.youtubeRegistrationRequestChecking
         .findUniqueOrThrow({ where: { id: p.data.id } })
-        .then((c) => YoutubeRegistrationRequestAcceptingModel.fromPrisma(c))
+        .then((c) => YoutubeRegistrationRequestAcceptingDTO.fromPrisma(c))
         .catch((e) => {
           logger.error({ error: e, path: info.path, id: p.data.id }, "Accepting not found");
           throw new GraphQLError("Something wrong happened");
