--- src/resolvers/Query/findYoutubeRegistrationRequest/findYoutubeRegistrationRequest.graphql
@@ -1,8 +1,0 @@
-type Query {
-  findYoutubeRegistrationRequest(input: FindYoutubeRegistrationRequestInput): YoutubeRegistrationRequest
-}
-
-input FindYoutubeRegistrationRequestInput {
-  id: ID
-  sourceId: String
-}
