--- src/resolvers/Query/getYoutubeRegistrationRequest/getYoutubeRegistrationRequest.graphql
@@ -1,3 +1,0 @@
-type Query {
-  getYoutubeRegistrationRequest(id: ID!): YoutubeRegistrationRequest!
-}
