--- src/resolvers/YoutubeRegistrationRequest/YoutubeRegistrationRequest.graphql
@@ -1,34 +1,0 @@
-type YoutubeRegistrationRequest {
-  id: ID!
-  title: String!
-
-  requestedBy: User!
-
-  sourceId: String!
-
-  title: String!
-  thumbnailUrl: String!
-
-  taggings: [YoutubeRegistrationRequestTagging!]!
-  semitaggings: [YoutubeRegistrationRequestSemitagging!]!
-
-  checked: Boolean!
-  checking: YoutubeRegistrationRequestChecking
-
-  originalUrl: String!
-  embedUrl: String!
-}
-
-type YoutubeRegistrationRequestTagging {
-  id: ID!
-  tag: Tag!
-  note: String
-}
-
-type YoutubeRegistrationRequestSemitagging {
-  id: ID!
-  name: String!
-  note: String
-}
-
-union YoutubeRegistrationRequestChecking = YoutubeRegistrationRequestAccepting | YoutubeRegistrationRequestRejecting
