--- src/resolvers/YoutubeRegistrationRequestAccepting/YoutubeRegistrationRequestAccepting.graphql
@@ -1,8 +1,0 @@
-type YoutubeRegistrationRequestAccepting {
-  request: YoutubeRegistrationRequest!
-
-  note: String
-  video: Video!
-
-  acceptedBy: User!
-}
