--- src/resolvers/YoutubeRegistrationRequestConnection/YoutubeRegistrationRequest.graphql
@@ -1,11 +1,0 @@
-type YoutubeRegistrationRequestConnection {
-  nodes: [YoutubeRegistrationRequest!]!
-  edges: [YoutubeRegistrationRequestEdge!]!
-  pageInfo: PageInfo!
-  totalCount: Int!
-}
-
-type YoutubeRegistrationRequestEdge {
-  cursor: String!
-  node: YoutubeRegistrationRequest!
-}
