--- src/resolvers/YoutubeRegistrationRequestRejecting/YoutubeRegistrationRequestRejecting.graphql
@@ -1,7 +1,0 @@
-type YoutubeRegistrationRequestRejecting {
-  request: YoutubeRegistrationRequest!
-
-  note: String
-
-  rejectedBy: User!
-}
