--- src/resolvers/YoutubeRegistrationRequestRejectingNotification/YoutubeRegistrationRequestRejectingNotification.graphql
@@ -1,8 +1,0 @@
-type YoutubeRegistrationRequestRejectingNotification implements Node & Notification {
-  id: ID!
-  createdAt: DateTime!
-  notifyTo: User!
-  watched: Boolean!
-
-  rejecting: YoutubeRegistrationRequestRejecting!
-}
